+++ baybe/surrogates/base.py
@@ -41,6 +41,8 @@ if TYPE_CHECKING:
     from botorch.posteriors import GPyTorchPosterior, Posterior
     from torch import Tensor
 
+    from baybe.surrogates.broadcasting import BroadcastingSurrogate
+
 _ONNX_ENCODING = "latin-1"
 """Constant signifying the encoding for onnx byte strings in pretrained models.
 
@@ -134,6 +136,22 @@ class Surrogate(ABC, SurrogateProtocol, SerialMixin):
         from baybe.surrogates._adapter import AdapterModel
 
         return AdapterModel(self)
+
+    def broadcast(self) -> BroadcastingSurrogate:
+        """Make the surrogate handle multiple targets via broadcasting.
+
+        If the surrogate only supports single targets, this method turns it into a
+        multi-target surrogate by replicating the model architecture for each observed
+        target. The resulting copies are trained independently, but share the same
+        architecture.
+
+        If the surrogate is itself already multi-target compatible, this operation
+        effectively disables the model's inherent multi-target mechanism by treating
+        it as a single-target surrogate and applying the same broadcasting mechanism.
+        """
+        from baybe.surrogates.broadcasting import BroadcastingSurrogate
+
+        return BroadcastingSurrogate(self)
 
     @staticmethod
     def _make_parameter_scaler_factory(